requirements.txt
distribute
Fabric
Flask
Flask-Cache
Flask-Classy
Flask-Script
Flask-Injector

raven
mailchimp
py_email_validation

dealer

unittest2
mock
nose
coverage

gevent
gunicorn